4R03 Replace old fiber/scheduler with core
