+++ examples/index.html
@@ -12,7 +12,7 @@
             <li><a href="counter.html">Counter</a></li>
             <li><a href="stopwatch.html">Stopwatch</a></li>
         </ul>
-        <p><a href="../test/core.html">Tests</a></p>
+        <p><a href="../test/">Tests</a></p>
     </body>
 </html>
