+++ lib/clock.js
@@ -1,16 +1,16 @@
 import { message } from "./util.js";
 
-const Playing = Symbol.for("playing");
-const Stopped = Symbol.for("stopped");
-
 export default class Clock {
     constructor() {
         this.currentTime = 0;
-        this.state = Stopped;
+        this.state = Clock.Stopped;
     }
 
+    static Stopped = Symbol.for("stopped");
+    static Playing = Symbol.for("playing");
+
     advance() {
-        if (this.state !== Playing || this.request) {
+        if (this.state !== Clock.Playing || this.request) {
             return;
         }
         this.request = window.requestAnimationFrame(() => { this.tick(); });
@@ -33,7 +33,7 @@ export default class Clock {
     }
 
     set now(end) {
-        console.assert(this.state === Stopped);
+        console.assert(this.state === Clock.Stopped);
         if (end !== this.currentTime) {
             const begin = this.currentTime;
             this.currentTime = end;
@@ -42,12 +42,12 @@ export default class Clock {
     }
 
     get now() {
-        return this.state === Stopped ? this.currentTime : Date.now() - this.startTime;
+        return this.state === Clock.Stopped ? this.currentTime : Date.now() - this.startTime;
     }
 
     start() {
-        if (this.state === Stopped) {
-            this.state = Playing;
+        if (this.state === Clock.Stopped) {
+            this.state = Clock.Playing;
             this.startTime = Date.now();
             this.lastUpdateTime = this.startTime;
             this.advance();
@@ -56,8 +56,8 @@ export default class Clock {
     }
 
     stop() {
-        if (this.state === Playing) {
-            this.state = Stopped;
+        if (this.state === Clock.Playing) {
+            this.state = Clock.Stopped;
             delete this.startTime;
             if (this.request) {
                 window.cancelAnimationFrame(this.request);
