+++ lib/fiber.js
@@ -1,585 +1,256 @@
-import { extend, isAsync, on, off, parseOffsetValue, typeOf } from "./util.js";
+import { extend, message, nop, remove } from "./util.js";
+
+const Ops = {
+    // Call f and handle its return value or error, then resume the fiber.
+    // Yield until the call ends. When going backward, handle the end of a ramp.
+    beginasync(scheduler, f, delegate) {
+        if (this.rate > 0) {
+            const instance = this.asyncDelegate = extend(delegate, { observedBegin: scheduler.now });
+            f(this, scheduler).then(value => {
+                if (this.asyncDelegate === instance) {
+                    delegate.asyncWillEnd?.call(delegate, value, this, scheduler);
+                    this.asyncDidEnd(scheduler);
+                }
+            }).catch(error => {
+                if (this.asyncDelegate === instance) {
+                    delegate.asyncWillEndWithError?.call(delegate, error, this, scheduler);
+                    this.asyncDidEnd(scheduler, error);
+                }
+            });
+            return Fiber.Yield;
+        }
+        // Reverse: the ramp ended, the async delegate can be deleted.
+        delete this.asyncDelegate;
+        scheduler.endRampForFiber(this, 0);
+    },
+
+    // When reversing an async call, simply ramp backward with the same
+    // duration. A reverse function f may have been provided by reverse().
+    endasync(scheduler, f) {
+        if (this.rate < 0) {
+            f?.(this, scheduler);
+            const [, begin] = this.trace.at(-1);
+            return scheduler.beginRampForFiber(this, this.now - begin, nop, 1);
+        }
+    },
+
+    // Increment (or decrement) the ever count of the fiber, reversing it when
+    // going backward.
+    ever(_, incr) {
+        this.ever += incr * (this.rate < 0 ? -1 : 1);
+        console.assert(this.ever >= 0);
+    },
+
+    // Create a new instance of the join delegate, storing the begin time and
+    // pending child fibers and calling the optional delegate method
+    // `fiberWillJoin` with the delegate as `this`. When there are no child
+    // fibers to join, end synchronously.
+    beginjoin(scheduler, delegate) {
+        if (this.rate > 0) {
+            if (!this.children) {
+                return;
+            }
+            console.assert(this.children.length > 0);
+            this.joinDelegate = extend(delegate, { pending: new Set(this.children) });
+            delegate.fiberWillJoin?.call(delegate, this, scheduler);
+            return Fiber.Yield;
+        }
+        // Reverse: the ramp ended, the join delegate can be deleted.
+        // fiberWillJoin may have had effects that we want to reverse.
+        delegate.fiberWillJoinReverse?.call(delegate, this, scheduler);
+        scheduler.endRampForFiber(this, 0);
+    },
+
+    // End a join: clean up and resume.
+    endjoin(scheduler) {
+        if (this.rate > 0) {
+            // The next op has already been pushed to the trace so save the
+            // children to the next to last entry, which is the endjoin.
+            console.assert(this.ops[this.trace.at(-2)[0]][0] === "endjoin");
+            this.trace.at(-2).push(this.children);
+            delete this.joinDelegate;
+            delete this.children;
+        } else {
+            const [, begin,, children] = this.trace.at(-1);
+            this.children = children;
+            scheduler.setChildFibersRate(this);
+            // Always yield, even when the join was synchronous; the fiber must
+            // then be scheduled if `beginRamp` does not do it.
+            if (!scheduler.beginRampForFiber(this, this.now - begin, nop, 1)) {
+                scheduler.scheduleFiber(this);
+            }
+            return Fiber.Yield;
+        }
+    },
+
+    // Begin a ramp, evaluating its effective duration if necessary, or end it
+    // when going backward.
+    beginramp(scheduler, dur, f) {
+        if (this.rate > 0) {
+            const effectiveDuration = Math.max(0, this.getEffectiveParameter(dur, scheduler));
+            return scheduler.beginRampForFiber(this, isNaN(effectiveDuration) ? 0 : effectiveDuration, f, 0);
+        } else {
+            // End back at the beginning of the ramp.
+            scheduler.endRampForFiber(this, 0);
+        }
+    },
+
+    // End a ramp (when going forward) or begin in reverse when going backward
+    // using the same duration as when going forward.
+    endramp(scheduler, f) {
+        if (this.rate > 0) {
+            scheduler.endRampForFiber(this, 1);
+        } else {
+            // Start a backward ramp with the same duration that just elapsed
+            // (or if the ramp was interrupted because the fiber was cancelled,
+            // with the same duration as the original ramp and picking up at
+            // the point where it was interrupted, which can be recalled from
+            // the trace).
+            const [, begin,, dur, p] = this.trace.at(-1);
+            return scheduler.beginRampForFiber(this, dur ?? this.now - begin, f, p ?? 1);
+        }
+    },
+
+    // Begin a repeat by setting up the delegate (if necessary) and calling its
+    // repeatShouldEnd() method with the current iteration count (starting at
+    // zero). If this returns true, then skip to after the repeat, otherwise
+    // attach an instance of the body fiber and yield.
+    beginrepeat(scheduler, delegate, body) {
+        if (this.rate > 0) {
+            if (!this.joinDelegate) {
+                this.joinDelegate = extend(delegate, { iteration: 0 });
+            } else {
+                this.joinDelegate.iteration += 1;
+            }
+            if (delegate.repeatShouldEnd?.call(delegate, this.joinDelegate.iteration, this, scheduler)) {
+                delete this.joinDelegate;
+                // Skip the next op.
+                return 1;
+            } else {
+                const instance = scheduler.attachFiber(this, extend(body, { id: Fiber.ID++ }));
+                this.joinDelegate.pending = new Set([instance]);
+                return Fiber.Yield;
+            }
+        }
+        // If there was an iteration, then a ramp began while the body is
+        // reversing.
+        if (scheduler.ramps.has(this)) {
+            scheduler.endRampForFiber(this, 0);
+        }
+    },
+
+    // End an iteration by going back to the beginning of the repeat.
+    endrepeat(scheduler) {
+        if (this.rate > 0) {
+            // Save the body instance so that it can be unended and skip back
+            // two ops to the matching beginrepeat.
+            console.assert(this.ops[this.trace.at(-2)[0]][0] === "endrepeat");
+            this.trace.at(-2).push(this.children);
+            delete this.children;
+            return -2;
+        }
+        // Similar to endjoin, unend the body fiber (stored above in the trace)
+        // and setup a ramp for the duration (always yielding) of that fiber.
+        return Ops.endjoin.call(this, scheduler);
+    },
+
+    // Attach an instance of a fiber to the parent.
+    spawn(scheduler, child) {
+        if (this.rate > 0) {
+            const instance = scheduler.attachFiber(this, extend(child, { id: Fiber.ID++ }));
+            this.trace.at(-2).push(instance);
+        } else {
+            const instance = this.trace.at(-1).at(-1);
+            console.assert(Object.getPrototypeOf(instance) === child);
+            console.assert(instance.parent === this);
+            delete instance.parent;
+            remove(this.children, instance);
+            if (this.children.length === 0) {
+                delete this.children;
+            }
+        }
+    },
+
+    // Call f synchronously with the fiber and scheduler as arguments.
+    sync(scheduler, f, g) {
+        if (this.rate > 0) {
+            f(this, scheduler);
+        } else {
+            console.assert(this.rate < 0);
+            g?.(this, scheduler);
+        }
+    }
+};
+
+const ReversibleArity = {
+    endasync: 1,
+    sync: 2
+};
 
 export default class Fiber {
-    static Names = new Map();
-    static Count = 0;
+    constructor() {
+        this.ops = [];
+        this.id = Fiber.ID++;
+    }
+
+    static ID = 0;
+    static Yield = Symbol.for("yield");
     static Cancelled = Symbol.for("cancelled");
 
-    constructor() {
-        this.id = Fiber.Count++;
-        this.ops = [];
-        this.metadata = [];
-    }
-
-    instantiate(...props) {
-        return extend(this, { id: `${Fiber.Count++}<${this.id}` }, ...props)
-    }
-
-    // Get the name of this fiber, if any.
-    get name() {
-        return Fiber.Names.get(this);
-    }
-
-    get value() {
-        if (this.result.error) {
-            return;
-        }
-        return this.result.value;
-    }
-
-    set value(value) {
-        delete this.result.error;
-        this.result.value = value;
-    }
-
-    get error() {
-        return this.result.error;
-    }
-
-    get handleResult() {
-        return (this.result.error && this.handleError.at(-1)) || (!this.result.error && this.handleValue.at(-1));
-    }
-
-    // True when the fiber is cancelled.
-    get isCancelled() {
-        return this.error === Fiber.Cancelled;
-    }
-
-    // Set the error field of the result and report it to the console.
-    errorWithMessage(error) {
-        this.result.error = error;
-        console.error(error.message ?? error);
-    }
-
-    // Convenience method to add new ops to the fiber, returning it for
-    // chaining.
-    op(op) {
-        this.ops.push(op);
+    // Add an async op to the fiber and return the fiber. A delegate can handle
+    // the result of the call with its optional `asyncWillEndWithValue` and
+    // `asyncWillEndWithError` methods that get called when the async call
+    // finishes and before moving to the next op.
+    async(f, delegate = {}) {
+        this.ops.push(["beginasync", f, delegate]);
+        this.ops.push(["endasync"]);
         return this;
     }
 
-    // Set metadata for the last operator.
-    meta(data) {
-        const i = this.ops.length - 1;
-        if (!this.metadata[i]) {
-            this.metadata[i] = data;
-        } else {
-            this.metadata[i] = extend(this.metadata[i], data);
-        }
+    // Wrap a block f (called with the fiber) into a pair of ever instructions,
+    // allowing the block to be executed even when an error occurs.
+    ever(f) {
+        this.ops.push(["ever", 1]);
+        f(this);
+        this.ops.push(["ever", -1]);
         return this;
     }
 
-    // Add custom undo behaviour to the last op if permitted.
-    undo(f) {
-        const i = this.ops.length - 1;
-        const data = this.metadata[i];
-        if (!data.undo) {
-            throw Error("Cannot set metadata for op");
-        }
-        this.metadata[i] = extend(data, { undo: f });
+    // Wait for child fibers to end, calling delegate methods before suspending
+    // and anytime a child fiber ends.
+    join(delegate = {}) {
+        this.ops.push(["beginjoin", delegate]);
+        this.ops.push(["endjoin"]);
         return this;
     }
 
-    // Get the effective duration of a delay or a ramp, which could be a
-    // number, a string, or a function returning a number or a string.
-    // Return nothing in case of error.
-    getEffectiveDuration(dur, scheduler) {
-        if (typeof dur === "function") {
-            try {
-                dur = dur(this, scheduler);
-            } catch (error) {
-                this.errorWithMessage(error);
-                return;
-            }
-        }
-        return typeof dur === "string" ? parseOffsetValue(dur) : dur;
-    }
-
-    // Evaluate a function given as argument with the fiber and scheduler,
-    // or return the value as is.
-    getEffectiveParameter(param, scheduler) {
-        try {
-            return this.getEffectiveParameterUnsafe(param, scheduler);
-        } catch (error) {
-            this.errorWithMessage(error);
-        }
-    }
-
-    // Unsafe version for use inside a try/catch.
-    getEffectiveParameterUnsafe(param, scheduler) {
-        return typeof param === "function" ? param(this, scheduler) : param;
-    }
-
-    // Call the function f with this fiber as a first parameter, passing all
-    // other additional parameters if needed.
-    macro(f, ...args) {
-        f(this, ...args);
-        return this;
-    }
-
-    // Set the name of the fiber so that it can be retrieved from the scheduler
-    // map of active fibers. It is an error to name the fiber with the same
-    // name as another running fiber.
-    // FIXME 4L01 Undo either
-    named(name) {
-        return this.op(function(scheduler) {
-            if (!this.handleResult) {
-                return;
-            }
-            try {
-                const effectiveName = this.getEffectiveParameterUnsafe(name, scheduler);
-                const previousName = this.name;
-                if (effectiveName === previousName) {
-                    return;
-                }
-                this.unops.push(previousName === undefined ?
-                    function(scheduler) {
-                        scheduler.removeNameForFiber(this);
-                    } : function(scheduler) {
-                        scheduler.setNameForFiber(this, previousName);
-                    }
-                );
-                if (effectiveName === undefined) {
-                    scheduler.removeNameForFiber(this);
-                } else {
-                    scheduler.setNameForFiber(this, effectiveName);
-                }
-            } catch (error) {
-                const value = this.result.value;
-                this.unops.push(function() {
-                    this.value = value;
-                });
-                this.errorWithMessage(error);
-            }
-        }).meta({ op: "named", name });
-    }
-
-    // Store the current value of the fiber under the given name in its scope.
-    store(name) {
-        return this.op(function(scheduler) {
-            if (!this.handleResult) {
-                return;
-            }
-            const effectiveName = this.getEffectiveParameter(name, scheduler);
-            if (effectiveName !== undefined) {
-                const shadowed = Object.hasOwn(this.scope, effectiveName);
-                const previous = this.scope[effectiveName];
-                this.unops.push(() => {
-                    if (shadowed) {
-                        this.scope[effectiveName] = previous;
-                    } else {
-                        delete this.scope[effectiveName];
-                    }
-                });
-                this.scope[effectiveName] = this.value;
-            }
-        }).meta({ op: "store", name });
-    }
-
-    // Call a function with this and the scheduler as parameters, and set the
-    // value of the fiber to the return value of that function. If the function
-    // is explicitly marked as being asynchronous, yield and resume once the
-    // function returns. Default undo is to revert to the previous value,
-    // treating an async call as a delay.
-    // FIXME 4I04 Cancelling async exec/effect
-    // FIXME 4L01 Undo either
-    exec(f) {
-        const i = this.ops.length;
-        return this.op(isAsync(f) ? function(scheduler) {
-            if (!this.handleResult) {
-                return;
-            }
-            const undo = this.metadata[i].undo;
-            f(this, scheduler).
-                then(v => {
-                    if (this.handleResult) {
-                        const value = this.value;
-                        const begin = this.now;
-                        this.value = v;
-                        this.now = scheduler.fiberLocalTime(this, scheduler.clock.now);
-                        const delay = begin - this.now;
-                        this.unops.push(typeof undo === "function" ? function(scheduler) {
-                            this.value = value;
-                            scheduler.setDelayForFiber(this, delay);
-                            undo(this, scheduler);
-                        } : function(scheduler) {
-                            this.value = value;
-                            scheduler.setDelayForFiber(this, delay);
-                        });
-                        scheduler.resumeFiber(this, scheduler.clock.now);
-                    }
-                }).
-                catch(error => {
-                    if (this.handleResult) {
-                        const value = this.value;
-                        this.errorWithMessage(error);
-                        const begin = this.now;
-                        this.value = value;
-                        this.now = scheduler.fiberLocalTime(this);
-                        const delay = begin - this.now;
-                        this.unops.push(typeof undo === "function" ? function(scheduler) {
-                            this.value = value;
-                            scheduler.setDelayForFiber(this, delay);
-                            undo(this, scheduler);
-                        } : function(scheduler) {
-                            this.value = value;
-                            scheduler.setDelayForFiber(this, delay);
-                        });
-                        scheduler.resumeFiber(this, scheduler.clock.now);
-                    }
-                });
-            this.yielded = true;
-        } : function(scheduler) {
-            if (!this.handleResult) {
-                return;
-            }
-            const value = this.value;
-            const undo = this.metadata[i].undo;
-            this.unops.push(typeof undo === "function" ?
-                function(scheduler) {
-                    this.value = value;
-                    undo(this, scheduler);
-                } : function() {
-                    this.value = value;
-                }
-            );
-            try {
-                this.value = f(this, scheduler);
-            } catch (error) {
-                this.errorWithMessage(error);
-            }
-        }).meta({ op: "exec", async: isAsync(f), undo: true });
-    }
-
-    // Effect calls the function `f` with the fiber and scheduler as arguments.
-    // `f` may be synchronous or asynchronous. The return value of `f` is
-    // discarded and the fiber value is unchanged (unless `f` mutates it as an
-    // effect).
-    // FIXME 4I04 Cancelling async exec/effect
-    effect(f) {
-        const i = this.ops.length;
-        return this.op(isAsync(f) ? function(scheduler) {
-            if (!this.handleResult) {
-                return;
-            }
-            f(this, scheduler).
-                catch(error => { this.errorWithMessage(error); }).
-                finally(() => {
-                    if (this.handleResult) {
-                        const begin = this.now;
-                        this.now = scheduler.fiberLocalTime(this, scheduler.clock.now);
-                        const delay = begin - this.now;
-                        const undo = this.metadata[i].undo;
-                        this.unops.push(typeof undo === "function" ? function(scheduler) {
-                            scheduler.setDelayForFiber(this, delay);
-                            undo(this, scheduler);
-                        } : function(scheduler) {
-                            scheduler.setDelayForFiber(this, delay);
-                        });
-                        scheduler.resumeFiber(this, scheduler.clock.now);
-                    }
-                });
-            this.yielded = true;
-        } : function(scheduler) {
-            if (!this.handleResult) {
-                return;
-            }
-            try {
-                f(this, scheduler);
-                const undo = this.metadata[i].undo;
-                if (typeof undo === "function") {
-                    this.unops.push(function(scheduler) {
-                        undo(this, scheduler);
-                    });
-                }
-            } catch (error) {
-                this.errorWithMessage(error);
-            }
-        }).meta({ op: "effect", async: isAsync(f), undo: true });
-    }
-
-    // Add an event listener and yield until the event occurs. Target and type
-    // may be functions that get evaluated with fiber and scheduler as
-    // parameters, and describe the target and the type of the event. By
-    // default, the fiber value is not affected. Target may be a DOM element,
-    // where `addEventListener` is used; otherwise, event expects a synchronous
-    // message and uses `on` for the event listener itself.
-    // When the event is received, the following delegate methods are called:
-    // * `eventShouldBeIgnored` allows ignoring an event (i.e., the fiber stays
-    //   suspended until an event is received and not ignore).
-    // * `eventWasHandled` allows calling functions like `preventDefault`
-    //   before the fiber resumes.
-    // Undo is simply a negative delay.
-    event(target, type, delegate = {}) {
-        const i = this.ops.length;
-        return this.op(function(scheduler) {
-            console.assert(!this.eventDelegate);
-            if (!this.handleResult) {
-                return;
-            }
-            const effectiveTarget = this.getEffectiveParameter(target, scheduler);
-            const effectiveType = this.getEffectiveParameter(type, scheduler);
-            if (!effectiveTarget || !effectiveType) {
-                return;
-            }
-            this.eventDelegate = extend(delegate, { target: effectiveTarget, type: effectiveType });
-            if (effectiveTarget.addEventListener) {
-                effectiveTarget.addEventListener(effectiveType, this.eventDelegate);
-                this.eventDelegate.handleEvent = event => {
-                    if (this.rate === 0 ||
-                        this.eventDelegate.eventShouldBeIgnored?.call(delegate, event, this, scheduler)) {
-                        return;
-                    }
-                    effectiveTarget.removeEventListener(effectiveType, this.eventDelegate);
-                    this.eventWasHandled(scheduler, delegate, event, i);
-                };
-            } else {
-                on(effectiveTarget, effectiveType, this.eventDelegate);
-                this.eventDelegate.handleMessage = message => {
-                    if (this.rate === 0 ||
-                        this.eventDelegate.eventShouldBeIgnored?.call(delegate, message, this, scheduler)) {
-                        return;
-                    }
-                    off(effectiveTarget, effectiveType, this.eventDelegate);
-                    this.eventWasHandled(scheduler, delegate, message, i);
-                };
-            }
-            this.yielded = true;
-        }).meta({ op: "event", target, type, undo: true });
-    }
-
-    // Event or message was handled: call the delegate method and resume the
-    // fiber.
-    eventWasHandled(scheduler, delegate, event, i) {
-        this.eventDelegate.eventWasHandled?.call(delegate, event, this, scheduler);
-        delete this.eventDelegate;
-        const begin = this.now;
-        this.now = scheduler.fiberLocalTime(this, scheduler.clock.now);
-        const delay = begin - this.now;
-        const undo = this.metadata[i].undo;
-        this.unops.push(typeof undo === "function" ?
-            function(scheduler) {
-                scheduler.setDelayForFiber(this, delay);
-                undo(this, scheduler);
-            } :
-            function(scheduler) {
-                scheduler.setDelayForFiber(this, delay);
-            }
+    // Add begin/end ramp ops to the fiber with the given duration and callback.
+    // Return the fiber.
+    ramp(dur, f = nop) {
+        this.ops.push(
+            ["beginramp", dur, f],
+            ["endramp", f]
         );
-        scheduler.resumeFiber(this, scheduler.clock.now);
-    }
-
-    // Normally fiber execution skips over errors, but either allows handling
-    // them by providing a path for values (f) and a path for errors (g). When
-    // the second path is omitted, both errors and values are handled; the
-    // fiber has an error if it is failing, and no error otherwise.
-    either(f, g) {
-        if (g) {
-            this.ops.push(function(scheduler) {
-                this.handleValue.push(true);
-                this.handleError.push(false);
-            });
-            this.meta({ op: "either/value" });
-            f(this);
-            this.ops.push(function(scheduler) {
-                const m = this.handleValue.length - 1;
-                console.assert(this.handleValue[m]);
-                this.handleValue[m] = false;
-                const n = this.handleError.length - 1;
-                console.assert(!this.handleError[n]);
-                this.handleError[n] = true;
-            });
-            this.meta({ op: "either/error" });
-            g(this);
-            this.ops.push(function(scheduler) {
-                console.assert(!this.handleValue.at(-1));
-                console.assert(this.handleError.at(-1));
-                this.handleValue.pop();
-                this.handleError.pop();
-            });
-            this.meta({ op: "either/end" });
-        } else {
-            this.ops.push(function() { this.handleError.push(true); });
-            this.meta({ op: "either/begin" });
-            f(this);
-            this.ops.push(function() { this.handleError.pop(); });
-            this.meta({ op: "either/end" });
-        }
         return this;
     }
 
-    // Spawn a single fiber executing f repeatedly. The delegate object allows
-    // controlling when the loop ends by being called before a new iteration
-    // is about to begin with the current iteration count (starting at 0 so
-    // that the whole repetition can be skipped altogether). Zero-duration
-    // repeats without a delegate cause an error to avoid infinite loops.
-    // The body of the loop runs as its own fiber.
+    // Repeat a block f an infinite number of times, or until the delegate
+    // method `repeatShouldEnd()` returns true.
     repeat(f, delegate = {}) {
-        const body = new Fiber().
-            op(function() {
-                // FIXME 4F01 Delegates should be instantiated
-                this.repeatDelegate = Object.create(delegate);
-                this.repeatDelegate.count = 0;
-            }).meta({ op: "repeat/init" }).
-            op(function(scheduler) {
-                if (!this.handleResult ||
-                    this.repeatDelegate.repeatShouldEnd?.call(delegate, this.repeatDelegate.count, this, scheduler)) {
-                    this.ip = this.ops.length - 1;
-                }
-                this.repeatDelegate.iterationTime = scheduler.now;
-            }).meta({ op: "repeat/enter" }).
-            macro(f).
-            op(function(scheduler) {
-                if (!this.repeatDelegate.repeatShouldEnd && scheduler.now === this.repeatDelegate.iterationTime) {
-                    this.errorWithMessage(Error("Zero duration repeat"));
-                } else {
-                    this.repeatDelegate.count += 1;
-                    this.ip = 1;
-                }
-            }).meta({ op: "repeat/loop" }).
-            op(function(scheduler) {
-                if (this.result) {
-                    if (this.error) {
-                        this.parent.result.error = this.error;
-                    } else {
-                        this.parent.value = this.value;
-                    }
-                }
-                delete this.repeatDelegate;
-                delete this.parent.child;
-                scheduler.resumeFiber(this.parent);
-            }).meta({ op: "repeat/leave" });
-        return this.op(function(scheduler) {
-            if (!this.handleResult) {
-                return;
-            }
-            this.child = body.instantiate({ parent: this });
-            scheduler.resetFiber(this.child);
-            scheduler.resumeFiber(this.child);
-            this.yielded = true;
-        }).meta({ op: "repeat/instantiate" });
-    }
-
-    // Repeat f over the values of the fiber.
-    each(f) {
-        const values = fiber => {
-            if (fiber.error) {
-                return [fiber.error];
-            }
-            const type = typeOf(fiber.value);
-            return type === "array" ? fiber.value :
-                type === "set" ? [...fiber.value.values()] :
-                type === "map" ? [...fiber.value.entries()] :
-                type === "object" ? [...Object.entries(fiber.value)] : [fiber.value];
-        };
-        return this.repeat(f, {
-            repeatShouldEnd(i, fiber) {
-                if (i === 0) {
-                    this.values = values(fiber);
-                }
-                if (i < this.values.length) {
-                    if (fiber.parent.error) {
-                        fiber.result.error = this.values[i];
-                    } else {
-                        fiber.value = this.values[i];
-                    }
-                } else {
-                    delete fiber.result;
-                    return true;
-                }
-            }
-        }, {}).meta({ op: "repeat/each" });
-    }
-
-    // Ramp from 0 to 1 over a given duration (dur may be a function or a fixed
-    // duration specified as a number of milliseconds or a SMIL clock value).
-    // The delegate method `rampDidProgress` gets called with three
-    // parameters `p`, `fiber` and `scheduler` (and `this` being the instance
-    // of the delegate object created for this ramp) on every update during the
-    // duration of the ramp. `rampDidProgress` is called first with p=0 at the
-    // exact time when the ramp starts and p=1 exactly when the ramp ends; in
-    // between, it may get called with 0 < p < 1 increasing at various times
-    // within the duration. When duration is infinite however, p is always 0.
-    // Undoing simply runs the ramp backward.
-    // FIXME 4E02 Count iterations for (infinite) ramps
-    ramp(dur, delegate) {
-        if (typeof dur === "function") {
-            this.ops.push(function(scheduler) {
-                if (!this.handleResult) {
-                    return;
-                }
-                const effectiveDur = this.getEffectiveDuration(dur, scheduler);
-                if (typeof effectiveDur === "number" && effectiveDur > 0) {
-                    scheduler.beginRampForFiber(this, effectiveDur, delegate);
-                    this.unops.push(function(scheduler) { scheduler.endRampForFiber(this); });
-                }
-            });
-            this.meta({ op: "ramp/begin/var" });
-            this.ops.push(function(scheduler) {
-                const dur = scheduler.endRampForFiber(this);
-                this.unops.push(function(scheduler) { scheduler.beginRampForFiber(this, -dur, delegate); });
-            });
-            this.meta({ op: "ramp/end" });
-        } else {
-            const effectiveDur = typeof dur === "string" ? parseOffsetValue(dur) : dur;
-            if (typeof dur === "number" && dur > 0) {
-                this.ops.push(function(scheduler) {
-                    scheduler.beginRampForFiber(this, effectiveDur, delegate);
-                    this.unops.push(function(scheduler) { scheduler.endRampForFiber(this); });
-                });
-                this.meta({ op: "ramp/begin" });
-                this.ops.push(function(scheduler) {
-                    scheduler.endRampForFiber(this);
-                    this.unops.push(function(scheduler) { scheduler.beginRampForFiber(this, -effectiveDur, delegate); });
-                });
-                this.meta({ op: "ramp/end" });
-            }
+        const body = new Fiber();
+        this.ops.push(["beginrepeat", delegate, body]);
+        this.ops.push(["endrepeat"]);
+        if (typeof f === "function") {
+            f(body);
+            return this;
         }
-        return this;
+        return body;
     }
 
-    // Wait for `dur`ms, unless the fiber is failing. If dur is a function,
-    // call it with `fiber` and `scheduler` as arguments and use the return
-    // value as the delay duration. The duration value may also be an offset
-    // value which gets parsed into a number of milliseconds. There is no
-    // effect if the final duration is not a number greater than zero. The
-    // value of the fiber is not affected by the delay.
-    delay(dur) {
-        return this.op(function(scheduler) {
-            if (!this.handleResult) {
-                return;
-            }
-            const effectiveDur = this.getEffectiveDuration(dur, scheduler);
-            if (typeof effectiveDur === "number" && effectiveDur > 0) {
-                scheduler.setDelayForFiber(this, effectiveDur);
-                this.unops.push(function(scheduler) {
-                    scheduler.setDelayForFiber(this, -effectiveDur);
-                    if (!this.yielded) {
-                        this.now -= effectiveDur;
-                    }
-                });
-                if (!this.yielded) {
-                    this.now += effectiveDur;
-                }
-            }
-        }).meta({ op: "delay", dur });
-    }
-
-    // Spawn a new fiber. The new fiber is created immediately as a child of
-    // this fiber and returned, unless a function is passed as the first
-    // argument, in which case that function is called with the child fiber
-    // as its only argument and the current fiber is returned. The child fiber
-    // will then be instantiated (since it may be spawned multiple times, for
-    // instance within a map) and begin after the parent fiber yields, in the
-    // same instant.
+    // Spawn a child fiber.
     spawn(f) {
         const child = new Fiber();
-        this.ops.push(function(scheduler) {
-            if (this.handleResult) {
-                scheduler.attachFiber(this, child.instantiate());
-            }
-        });
-        this.meta({ op: "spawn" });
+        this.ops.push(["spawn", child]);
         if (typeof f === "function") {
             f(child);
             return this;
@@ -587,144 +258,138 @@ export default class Fiber {
         return child;
     }
 
-    // Map spawns a new fiber for each item in the fiber value.
-    map(f) {
-        const template = new Fiber();
-        f(template);
-        return this.op(function(scheduler) {
-            if (!this.handleResult) {
-                return;
-            }
-            const type = this.error ? "error" : typeOf(this.value);
-            const values =
-                type === "set" ? this.value.values() :
-                type === "map" ? this.value.entries() :
-                type === "object" ? Object.entries(this.value) : this.value;
-            if (type === "array" || type === "set" || type === "map" || type === "object") {
-                for (const value of values) {
-                    scheduler.attachFiber(this, template.instantiate({ result: { value } }));
+    // Add a sync op to the fiber and return it.
+    sync(f) {
+        this.ops.push(["sync", f]);
+        return this;
+    }
+
+    // Provide a custom reverse effect to ops that allow it.
+    reverse(f) {
+        if (this.ops.length === 0) {
+            throw Error("Nothing to reverse");
+        }
+        const op = this.ops.at(-1);
+        const arity = ReversibleArity[op[0]];
+        if (isNaN(arity)) {
+            throw Error(`Cannot provide a reverse effect for ${op[0]}`);
+        }
+        if (op.length > arity) {
+            throw Error(`Already provided a reverse effect for ${op[0]}`);
+        }
+        op.push(f);
+        return this;
+    }
+
+    // Run all ops, building a trace of the execution of instructions. When
+    // going forward, the last element in the trace gives the index of the
+    // next instruction to execute as well as the current local time. Executing
+    // an instruction returns true when the fiber should yield. The generator
+    // returns when reaching the end of the sequence, or when an error occurs.
+    *run(scheduler) {
+        if (!Object.hasOwn(this, "now")) {
+            this.now = 0;
+            this.trace = [[0, this.now]];
+            this.spawns = new Map();
+            this.ever = 0;
+            this.rate = this.parent?.rate ?? 1;
+            this.ownRate = 1;
+            this.scope = this.parent ? Object.create(this.parent.scope) : {};
+            this.parent?.childFiberWillBegin(this);
+        }
+        if (this.rate === 0) {
+            yield;
+        }
+        for (
+            let n = this.ops.length, [ip] = this.trace.at(-1);
+            ((this.rate > 0 && 0 <= ip && ip < n) || (this.rate < 0 && this.trace.length > 1));
+        ) {
+            let allowError = this.ever > 0;
+            if (this.rate < 0) {
+                // FIXME 4M08 Core: redo
+                const [nextip, end, error] = this.trace.pop();
+                ip = nextip - 1;
+                this.error = error;
+                if (error && !allowError) {
+                    // If we are recovering from an async error or from a
+                    // cancelled ramp, we still need to ramp back.
+                    const [_, begin, unerror] = this.trace.at(-1);
+                    if (begin < end && error !== unerror) {
+                        console.assert(!unerror);
+                        allowError = true;
+                    }
                 }
-            } else {
-                scheduler.attachFiber(this, template);
             }
-        }).meta({ op: "map" });
-    }
-
-    // Yield and wait for child fibers to end. If there are no child fibers,
-    // or the fiber is failing, do nothing.
-    join(delegate = {}) {
-        return this.op(function(scheduler) {
-            if (!this.handleResult) {
-                // FIXME what to do about children?
-                if (this.children?.length > 0) {
-                    console.warn(`Not joinining because of error; pending children: ${this.children.size}`);
+            const [op, ...args] = this.ops[ip];
+            if (this.rate > 0) {
+                ip += 1;
+                this.trace.push([ip, this.now, this.error]);
+            }
+            if (this.error && !allowError && op !== "ever") {
+                continue;
+            }
+            try {
+                const v = Ops[op].call(this, scheduler, ...args);
+                if (v === Fiber.Yield || this.rate === 0) {
+                    yield;
+                } else if (typeof v === "number") {
+                    ip += v;
                 }
-                return;
-            }
-            console.assert(!this.joinDelegate);
-            if (this.children) {
-                console.assert(this.children.length > 0);
-                this.joinDelegate = extend(delegate, { pending: new Set(this.children) });
-                this.joinDelegate?.fiberWillJoin?.call(this.joinDelegate, this, scheduler);
-                // FIXME 4J0E Force join
-                if (this.children.length === 0) {
-                    delete this.joinDelegate;
-                } else {
-                    this.yielded = true;
-                }
-            }
-        }).meta({ op: "join" });
-    }
-}
-
-// Convenience join delegate function to cancel all remaining siblings of a
-// fiber that ended.
-export function cancelSiblings(delegate, scheduler) {
-    const siblings = [...delegate.pending];
-    delegate.pending.clear();
-    for (const sibling of siblings) {
-        scheduler.cancelFiber(sibling);
-    }
-}
-
-// Delegate to collect all child fiber values in order. Fails immediately with
-// the error of a child fiber that fails.
-export const All = {
-    fiberWillJoin(fiber) {
-        this.values = new Array(fiber.children.length).fill();
-    },
-
-    childFiberDidEnd(child, scheduler) {
-        const fiber = child.parent;
-        if (child.error) {
-            cancelSiblings(this, scheduler);
-            fiber.result.error = child.error;
-            delete this.values;
-        } else {
-            this.values[fiber.children.indexOf(child)] = child.value;
-            if (this.pending.size === 0) {
-                fiber.value = this.values;
+            } catch (error) {
+                this.errorWithMessage(scheduler, error);
             }
         }
     }
-};
 
-// Delegate to collect all child fiber values in the order in which they
-// finished. Fails immediately with the error of a child fiber that fails.
-export const Last = {
-    fiberWillJoin(fiber) {
-        this.values = [];
-    },
-
-    childFiberDidEnd(child, scheduler) {
-        const fiber = child.parent;
-        if (child.error) {
-            cancelSiblings(this, scheduler);
-            fiber.result.error = child.error;
-            delete this.values;
-        } else {
-            this.values.push(child.value);
-            if (this.pending.size === 0) {
-                child.parent.value = this.values;
-            }
+    // Resume after an async call ended, updating the fiber local time.
+    asyncDidEnd(scheduler, error) {
+        console.assert(this.rate >= 0);
+        const now = scheduler.clock.now;
+        this.now += (now - this.asyncDelegate.observedBegin) * this.rate;
+        delete this.asyncDelegate;
+        if (this.rate > 0) {
+            scheduler.scheduleFiber(this, now);
         }
+        if (error) {
+            this.errorWithMessage(scheduler, error);
+        }
+        return true;
     }
-};
 
-// Delegate to cancel all siblings once the first fiber ends. Skip errors,
-// unless all fibers do fail in which case the last error is reported.
-export const Gate = {
-    childFiberDidEnd(child, scheduler) {
-        if (child.error) {
-            if (this.pending.size === 0) {
-                child.parent.result.error = child.error;
-            }
-        } else {
-            cancelSiblings(this, scheduler);
-            return true;
-        }
+    // When a fiber begins, set its spawn time to keep track of its local time.
+    childFiberWillBegin(fiber) {
+        console.assert(!this.spawns.has(fiber));
+        this.spawns.set(fiber, this.now);
     }
-};
 
-// Same as Gate, but also sets the value of the fiber to the value of the first
-// fiber that ended with a value.
-export const First = {
-    childFiberDidEnd(child, scheduler) {
-        if (Gate.childFiberDidEnd.call(this, child, scheduler)) {
-            child.parent.value = child.value;
+    // Handle child fibers ending when a join is in progress. Return true when
+    // all child fibers have ended.
+    childFiberDidEnd(fiber, scheduler) {
+        if (!this.joinDelegate) {
+            return;
         }
+        const end = this.spawns.get(fiber) + fiber.now;
+        console.assert(end >= this.now);
+        this.now = end;
+        this.spawns.delete(fiber);
+        console.assert(this.joinDelegate.pending.has(fiber));
+        this.joinDelegate.pending.delete(fiber);
+        const delegate = Object.getPrototypeOf(this.joinDelegate);
+        delegate.childFiberDidJoin?.call(delegate, fiber, scheduler);
+        return this.joinDelegate.pending.size === 0;
     }
-};
 
-// Join a single fiber and use its eventual error or value.
-export const Single = {
-    childFiberDidEnd(child, scheduler) {
-        console.assert(this.pending.size === 0);
-        const fiber = child.parent;
-        if (child.error) {
-            fiber.result.error = child.error;
-        } else {
-            fiber.value = child.value;
+    // Set the error field of the result and send a message to report it.
+    errorWithMessage(scheduler, error) {
+        this.error = error;
+        message(scheduler, "error", { fiber: this, error });
+    }
+
+    // Evaluate a parameter if necessary, or pass its value through.
+    getEffectiveParameter(value, scheduler) {
+        if (typeof value === "function") {
+            return value(this, scheduler);
         }
+        return value;
     }
 }
