+++ lib/prelude.js
@@ -1,4 +1,5 @@
-import { Fiber, Scheduler } from "./core.js";
+import Fiber from "./fiber.js";
+import Scheduler from "./scheduler.js";
 export { Fiber, Scheduler };
 
 // Event is an async command that waits for a specific event before continuing.
