+++ lib/scheduler.js
@@ -1,228 +1,35 @@
-import { remove, PriorityQueue, on, message } from "./util.js";
+import { message, on, PriorityQueue, remove } from "./util.js";
 
 import Clock from "./clock.js";
 import Fiber from "./fiber.js";
 
 export default class Scheduler {
-    // Create a new scheduler with a default clock. It is generally more
-    // convenient to use init() or run() rather than creating a scheduler on
-    // its own.
-    // FIXME 4H0B Scheduler.init() and Scheduler.run()
     constructor() {
+        // Create a default clock to drive the updates.
         this.clock = new Clock();
         on(this.clock, "tick", ({ begin, end }) => { this.update(begin, end); });
+
+        // All instants (times) at which fibers are scheduled.
         this.instants = new PriorityQueue();
-        this.fibers = new Set();
-        this.fibersByName = new Map();
+
+        // Lists of fibers scheduled at a given instant.
         this.fibersByInstant = new Map();
         this.instantsByFiber = new Map();
-        this.delays = new Map();
+
+        // Keep track of active fibers, mapping the run generator to the fiber.
+        this.fibers = new Map();
+
+        // List of fibers that have an ongoing ramp.
         this.ramps = new Map();
+
+        // Scheduler time.
         this.now = 0;
     }
 
-    // Create a new scheduler and a main fiber, then start the clock. The fiber
-    // is returned so that new instructions can be added immediately.
-    static run() {
-        const scheduler = new Scheduler();
-        const fiber = new Fiber();
-        scheduler.clock.start();
-        scheduler.resetFiber(fiber);
-        scheduler.resumeFiber(fiber);
-        return fiber;
-    }
-
-    // FIXME 4H0B Scheduler.init() and Scheduler.run()
-    static initWithFiber(fiber) {
-        const scheduler = new Scheduler();
-        scheduler.resetFiber(fiber);
-        scheduler.resumeFiber(fiber);
-        return scheduler;
-    }
-
-    // Retrieve a fiber by its name.
-    fiberNamed(name) {
-        return this.fibersByName.get(name);
-    }
-
-    // Get the local time of an active fiber.
-    fiberLocalTime(fiber, now) {
-        return fiber.now + (
-            isFinite(fiber.rate) ? ((now ?? this.now) - (fiber.lastUpdateTime ?? fiber.beginTime)) * fiber.rate : 0
-        );
-    }
-
-    // Resume a fiber, as soon as possible, or at time t in the future. When
-    // resuming now, add to the resume queue of the current update loop; add
-    // at the end when the deferred flag is set (used for joining).
-    resumeFiber(fiber, t, deferred = false) {
-        if (fiber.rate === 0) {
-            return;
-        }
-        t = t ?? this.now;
-        console.assert(t >= this.now);
-        console.assert(!this.instantsByFiber.has(fiber));
-        this.instantsByFiber.set(fiber, t);
-        if (t === this.now && this.resumeQueues) {
-            this.resumeQueues[deferred ? 1 : 0].push(fiber);
-            return;
-        }
-        if (!this.fibersByInstant.has(t)) {
-            this.instants.insert(t);
-            this.fibersByInstant.set(t, []);
-        }
-        this.fibersByInstant.get(t).push(fiber);
-        this.clock.advance();
-    }
-
-    // Reschedule a fiber that is currently yielding to a new time t.
-    rescheduleFiber(fiber, t) {
-        if (this.instantsByFiber.has(fiber)) {
-            t = t ?? this.now;
-            if (t === this.instantsByFiber.get(fiber)) {
-                return;
-            }
-            remove(this.fibersByInstant.get(this.instantsByFiber.get(fiber)), fiber);
-            this.instantsByFiber.delete(fiber);
-        }
-        this.resumeFiber(fiber, t);
-    }
-
-    // Delay a fiber by `dur` which is expected to be greater than zero. The
-    // fiber yields and is scheduled to resume after taking into account the
-    // duration of the delay and the rate of the fiber.
-    setDelayForFiber(fiber, dur) {
-        console.assert(!this.delays.has(fiber) && !this.ramps.has(fiber));
-        const begin = this.now;
-        const rate = fiber.rate;
-        const effectiveDur = dur / rate;
-        if (effectiveDur > 0) {
-            this.delays.set(fiber, { begin, dur, rate, fiberBegin: fiber.now, fiberEnd: fiber.now + dur });
-            fiber.yielded = true;
-            this.resumeFiber(fiber, begin + effectiveDur);
-        }
-    }
-
-    // Begin a ramp for `dur` and call the `rampDidProgress` delegate with
-    // p = 0. The fiber yields and is scheduled to resume after the duration
-    // of the ramp. When undoing, begin at p = 1 and ramp back down to 0.
-    beginRampForFiber(fiber, dur, delegate) {
-        if (!fiber.handleResult) {
-            return;
-        }
-        console.assert(!this.delays.has(fiber) && !this.ramps.has(fiber));
-        const begin = this.now;
-        delegate = Object.create(delegate ?? {});
-        const rate = fiber.rate;
-        delegate.rampDidProgress?.call(delegate, rate < 0 ? 1 : 0, fiber, this);
-        this.ramps.set(fiber, { delegate, begin, dur, rate, fiberBegin: fiber.now, fiberEnd: fiber.now + dur });
-        fiber.yielded = true;
-        this.resumeFiber(fiber, begin + dur / rate);
-    }
-
-    // End the ramp when the fiber resumes and call the `rampDidProgress`
-    // delegate with p = 1. When undoing, end at 0.
-    endRampForFiber(fiber) {
-        if (!fiber.handleResult) {
-            return;
-        }
-        const { delegate, fiberBegin, fiberEnd, dur } = this.ramps.get(fiber);
-        this.ramps.delete(fiber);
-        fiber.now = fiberEnd;
-        delegate.rampDidProgress?.call(delegate, fiber.rate < 0 ? 0 : 1, fiber, this);
-        return dur;
-    }
-
-    // Reset the fiber by setting its begin time to the current time of the
-    // sechduler, its ip to 0, and initializing various state from the parent
-    // fiber.
-    resetFiber(fiber) {
-        console.assert(!this.fibers.has(fiber));
-        this.fibers.add(fiber);
-        fiber.beginTime = this.now;
-        fiber.now = 0;
-        delete fiber.endTime;
-        fiber.rate = fiber.parent?.rate ?? 1;
-        fiber.ownRate = 1;
-        fiber.ip = 0;
-        fiber.unops = [];
-        fiber.handleValue = [fiber.parent?.handleValue.at(-1) ?? true];
-        fiber.handleError = [fiber.parent?.handleError.at(-1) ?? false];
-        fiber.scope = fiber.parent ? Object.create(fiber.parent.scope) : {};
-        if (!Object.hasOwn(fiber, "result")) {
-            fiber.result = fiber.handleError.at(-1) && fiber.parent?.error ?
-                { error: fiber.parent.error } :
-                { value: fiber.parent?.value };
-        }
-    }
-
-    // Register a name for the fiber.
-    setNameForFiber(fiber, name) {
-        console.assert(name !== undefined);
-        if (this.fibersByName.has(name)) {
-            throw Error("A fiber with the same name is already running");
-        }
-        this.removeNameForFiber(fiber);
-        this.fibersByName.set(name, fiber);
-        Fiber.Names.set(fiber, name);
-    }
-
-    // Remove the name of the fiber, if any;
-    removeNameForFiber(fiber) {
-        const name = fiber.name;
-        if (this.fibersByName.has(name)) {
-            console.assert(this.fibersByName.get(name) === fiber);
-            this.fibersByName.delete(name);
-            Fiber.Names.delete(fiber);
-        }
-    }
-
-    // Cancel a fiber and its pending children, if joining. This sets its error
-    // value to Cancelled and resumes immediately, leaving the fiber a chance
-    // to handle cancellation gracefully. The current value is not overwritten.
-    // FIXME 4J0C Cancel a fiber with handleError=true
-    cancelFiber(fiber) {
-        console.assert(!fiber.endTime);
-        fiber.result.error = Fiber.Cancelled;
-        fiber.parent?.children?.splice(fiber.parent?.children?.indexOf(fiber), 1);
-        if (fiber.handleError.at(-1)) {
-            if (fiber.joinDelegate) {
-                for (const child of fiber.joinDelegate.pending) {
-                    this.cancelFiber(child);
-                }
-            }
-            return;
-        }
-        if (fiber.eventDelegate) {
-            // The fiber was waiting for an event.
-            const { target, type } = fiber.eventDelegate;
-            if (target.addEventListener) {
-                target.removeEventListener(type, fiber.eventDelegate);
-            } else {
-                off(target, type, fiber.eventDelegate);
-            }
-            delete fiber.eventDelegate;
-        }
-        if (fiber.joinDelegate) {
-            // The fiber is joining; cancel the pending children.
-            // Do not cancel children when not joining.
-            const pending = fiber.joinDelegate.pending;
-            delete fiber.joinDelegate;
-            for (const child of pending) {
-                this.cancelFiber(child);
-            }
-        }
-        if (fiber.child) {
-            // The fiber is joining a single fiber (as used by repeat).
-            this.cancelFiber(fiber.child);
-            delete fiber.child;
-        } else if (fiber.yielded) {
-            this.rescheduleFiber(fiber);
-        }
-    }
-
-    // Attach a child fiber like spawn does, creating it if necessary. This can
-    // be used to add new child fibers while the parent is still joining.
+    // Attach a child fiber (creating it if necessary). This can be used to add
+    // new child fibers while the parent is still joining, or based on dynamic
+    // input. The fiber is added to the spawn queue (if spawned during an update
+    // loop) or scheduled at the current time.
     attachFiber(fiber, child) {
         if (!child) {
             child = new Fiber();
@@ -232,148 +39,240 @@ export default class Scheduler {
             fiber.children = [];
         }
         fiber.children.push(child);
-        this.resetFiber(child);
-        this.resumeFiber(child);
+        if (this.spawnQueue) {
+            this.spawnQueue.push(child);
+        } else {
+            this.scheduleFiber(child, this.clock.now);
+        }
         return child;
     }
 
-    // A fiber ended. Its end time is set and its parent (if any) is notified.
-    // When a child fiber ends, remove it from the set of pending children
-    // and resume when the set becomes empty. Do nothing if the child is not
-    // pending (e.g., the fiber is not actually joining).
+    // Schedule a fiber to resume execution at time t (now by default). When t
+    // is infinite, the fiber is not actually added to the schedule.
+    scheduleFiber(fiber, t) {
+        t = t ?? this.now;
+        console.assert(t >= this.now);
+        this.instantsByFiber.set(fiber, t);
+        if (!isFinite(t)) {
+            return;
+        }
+        if (!this.fibersByInstant.has(t)) {
+            this.instants.insert(t);
+            this.fibersByInstant.set(t, [fiber]);
+        } else {
+            this.fibersByInstant.get(t).push(fiber);
+        }
+        this.clock.advance();
+    }
+
+    // Reschedule a fiber by removing it from the schedule and resuming it
+    // again.
+    rescheduleFiber(fiber, t) {
+        const instant = this.instantsByFiber.get(fiber);
+        if (instant === t) {
+            return;
+        }
+        if (isFinite(instant)) {
+            remove(this.fibersByInstant.get(instant), fiber);
+        }
+        this.instantsByFiber.delete(fiber);
+        this.scheduleFiber(fiber, t);
+    }
+
+    // Cancel a fiber by setting its error to the special Cancelled value.
+    // Async calls are cancelled then the fiber resumes—unless within an
+    // ever() block. Finally, children are cancelled.
+    cancelFiber(fiber) {
+        console.assert(fiber.rate > 0);
+        console.assert(!Object.hasOwn(fiber, "observedEnd"));
+        fiber.error = Fiber.Cancelled;
+        if (fiber.ever === 0) {
+            if (fiber.asyncDelegate) {
+                fiber.now = (this.now - fiber.asyncDelegate.observedBegin) * fiber.rate;
+                delete fiber.asyncDelegate;
+                this.scheduleFiber(fiber);
+            } else {
+                const { begin, dur, observedBegin, observedDur } = this.ramps.get(fiber);
+                this.ramps.delete(fiber);
+                const p = (this.now - observedBegin) / observedDur;
+                fiber.now = begin + p * dur;
+                fiber.trace.at(-1).push(dur, p);
+                this.rescheduleFiber(fiber, this.now);
+            }
+        }
+        if (this.children) {
+            for (const child of this.children) {
+                this.cancelFiber(child);
+            }
+        }
+    }
+
+    // The fiber is not active anymore. Record the observed end time for
+    // reversing spawns. Notify the parent since it may be joining, and as a
+    // result need to be scheduled when the last child has ended.
     fiberEnded(fiber) {
-        console.assert(this.fibers.has(fiber));
+        fiber.observedEnd = this.now;
         this.fibers.delete(fiber);
-        this.removeNameForFiber(fiber);
-        console.assert((fiber.rate > 0 && fiber.ip === fiber.ops.length) || (fiber.rate < 0 && fiber.unip === 0));
-        delete fiber.ip;
-        delete fiber.unip;
-        delete fiber.unops;
-        delete fiber.now;
+        if (fiber.parent?.childFiberDidEnd(fiber, this)) {
+            this.joinQueue.push(fiber.parent);
+        }
+    }
+
+    // Set the rate of the fiber, which is affected by its parent’s (if any),
+    // and then set the rate of the child fibers.
+    // If a ramp is ongoing, update its observed begin and duration to reflect
+    // the new rate and reschedule to fiber with the new observed end.
+    setFiberRate(fiber, rate) {
+        if (rate === fiber.ownRate) {
+            return;
+        }
+        const effectiveRate = rate * (fiber.parent?.rate ?? 1);
+        fiber.ownRate = rate;
+        this.setRateForSingleFiber(fiber, effectiveRate);
+        this.setChildFibersRate(fiber);
+    }
+
+    // Once the fiber rate has been set, propagate the change to child fibers.
+    // This is also used when reversing a join.
+    setChildFibersRate(fiber) {
+        if (!fiber.children) {
+            return;
+        }
+        // Set the rate of all descendants in depth-first order. We then set
+        // the descendants rate, and later unend them (in reverse order) if
+        // reversing.
+        const queue = fiber.children.slice();
+        const fibers = [];
+        while (queue.length > 0) {
+            const fiber = queue.shift();
+            this.setRateForSingleFiber(fiber, fiber.parent.rate * fiber.ownRate);
+            if (fiber.observedEnd >= 0) {
+                fibers.unshift(fiber);
+            }
+            if (fiber.children) {
+                for (let i = fiber.children.length - 1; i >= 0; --i) {
+                    queue.unshift(fiber.children[i]);
+                }
+            }
+        }
         if (fiber.rate < 0) {
-            // FIXME unspawn
-            return;
-        }
-        fiber.endTime = this.now;
-        const parent = fiber.parent;
-        if (!parent?.joinDelegate?.pending.has(fiber)) {
-            return;
-        }
-        parent.joinDelegate.pending.delete(fiber);
-        parent.joinDelegate.childFiberDidEnd?.call(parent.joinDelegate, fiber, this);
-        if (parent.joinDelegate.pending.size === 0) {
-            delete parent.children;
-            delete parent.joinDelegate;
-            parent.now = this.fiberLocalTime(parent);
-            this.resumeFiber(parent, this.now, true);
+            for (const fiber of fibers) {
+                // The unend time is the elapsed time between the observed
+                // end and the current scheduler time, but should be
+                // adjusted by rate.
+                const unend = this.now - fiber.observedEnd;
+                this.scheduleFiber(fiber, this.now - unend / fiber.rate);
+            }
         }
     }
 
-    // Update the current delay or ramp duration for the fiber, if any. If the
-    // new end time of the delay or ramp is before the current time, reschedule
-    // the fiber now.
-    updateDelayForFiber(fiber, dur) {
-        const update = delay => {
-            delay.dur = dur;
-            delay.fiberEnd = Math.max(fiber.now + dur, this.fiberLocalTime(fiber));
-            console.assert(fiber.rate === delay.rate);
-            this.rescheduleFiber(fiber, Math.max(delay.begin + dur / fiber.rate, this.now));
-        };
-        if (this.delays.has(fiber)) {
-            update(this.delays.get(fiber));
-        } else if (this.ramps.has(fiber)) {
-            const delay = this.ramps.get(fiber);
-            fiber.now = delay.fiberBegin;
-            update(delay);
-        }
-    }
-
-    // Set a new rate for the fiber. If it has a current delay or ramp, update
-    // its duration as well to reflect the change. Also adjust the begin time
-    // of the ramp to reflect the change of rate from this point on. Set the
-    // rate for the child fiber as well by multiplying the child’s own rate
-    // with the new rate; do not set the own rate of the children though (so
-    // that they can resume at the right rate after pausing, for instance).
-    setRateForFiber(fiber, rate, setOwnRate = true) {
+    // Only set the rate for the fiber, updating ramp and async timing as
+    // necessary. This is called by setFiberRate for itself and its descendants.
+    setRateForSingleFiber(fiber, rate) {
         if (rate === fiber.rate) {
             return;
         }
-        // FIXME 4L0D Reverse a reversed fiber
-        const reverse = fiber.rate >= 0 && rate < 0;
-        if (reverse) {
-            fiber.unip = fiber.unops.length;
-        }
-        if (setOwnRate) {
-            fiber.ownRate = rate;
+        if (this.ramps.has(fiber)) {
+            // Update the observed times of the current ramp with the new rate.
+            const ramp = this.ramps.get(fiber);
+            const elapsed = ramp.elapsed ?? (this.now - ramp.observedBegin) * fiber.rate;
+            delete ramp.elapsed;
+            if (rate === 0) {
+                // Pause the ramp, recording the elapsed time for resuming
+                // later.
+                ramp.elapsed = elapsed;
+                this.rescheduleFiber(fiber, Infinity);
+            } else if (isFinite(ramp.dur) && isFinite(rate)) {
+                ramp.observedBegin = this.now - elapsed / rate;
+                ramp.observedDur = ramp.dur / rate;
+                ramp.observedEnd = ramp.observedBegin + ramp.observedDur;
+                const end = rate > 0 ? ramp.observedEnd : ramp.observedBegin;
+                // If the ramp duration changed when the fiber was paused, it
+                // may have already ended by now.
+                this.rescheduleFiber(fiber, Math.max(end, this.now));
+            } else if (rate < 0) {
+                // Infinite ramp now has a finite observed duration based on the elapsed time.
+                ramp.observedBegin = this.now - elapsed / rate;
+                this.rescheduleFiber(fiber, ramp.observedBegin);
+            } else {
+                ramp.observedBegin = this.now - elapsed / rate;
+                this.rescheduleFiber(fiber, Infinity);
+            }
+        } else if (fiber.asyncDelegate && fiber.rate >= 0 && rate < 0) {
+            // Reversing an async call.
+            const now = this.clock.now;
+            fiber.now += (now - fiber.asyncDelegate.observedBegin) * fiber.rate;
+            fiber.trace.push([fiber.trace.at(-1)[0] + 1, fiber.now]);
+            this.scheduleFiber(fiber);
+        } else if (fiber.rate === 0) {
+            // Resume the fiber, for which time has not passed.
+            this.scheduleFiber(fiber);
         }
         fiber.rate = rate;
+    }
 
-        const update = delay => {
-            if (rate === 0) {
-                delay.p = (this.now - delay.begin) / (delay.dur / delay.rate);
-                remove(this.fibersByInstant.get(this.instantsByFiber.get(fiber)), fiber);
-                this.instantsByFiber.delete(fiber);
-            } else {
-                const elapsed = this.now - delay.begin;
-                if (reverse) {
-                    this.rescheduleFiber(fiber, this.now - elapsed / rate);
-                    const fiberBegin = delay.fiberBegin;
-                    delay.fiberBegin = delay.fiberEnd;
-                    delay.fiberEnd = fiberBegin;
-                } else {
-                    const p = delay.p ?? elapsed / (delay.dur / delay.rate);
-                    const dur = delay.dur / rate;
-                    delete delay.p;
-                    delay.begin = this.now - p * dur;
-                    this.rescheduleFiber(fiber, this.now + (1 - p) * dur);
-                }
-            }
-            delay.rate = rate;
-        };
-
-        if (this.delays.has(fiber)) {
-            update(this.delays.get(fiber));
-            if (reverse) {
-                fiber.unip -= 1;
-            }
-        } else if (this.ramps.has(fiber)) {
-            update(this.ramps.get(fiber));
-        } else if (rate === 0) {
-            console.assert(!this.instantsByFiber.has(fiber));
-            if (this.currentFiber === fiber) {
-                // The fiber is setting the rate to zero itself, so create
-                // a dummy zero-duration delay for resuming it when the rate
-                // becomes non-zero again.
-                this.delays.set(fiber, { begin: this.now, dur: 0, rate: 0, p: 0, fiberEnd: fiber.now });
-            }
-            fiber.yielded = true;
-        }
-        if (fiber.children) {
-            for (const child of fiber.children) {
-                this.setRateForFiber(child, rate * child.ownRate, false);
-            }
+    // Add a ramp object for this fiber and call its update function with a
+    // starting progress value p (0 when going forward, 0 < p ≤ 1 when going
+    // backward). Return Yield if the fiber needs to yield. A ramp stores its
+    // local begin time and duration; also its observed begin time, end
+    // time and duration, and the progress callback f. The observed duration is
+    // negative when the ramp goes backward (and observedBegin > observedEnd).
+    beginRampForFiber(fiber, dur, f, p) {
+        console.assert(!this.ramps.has(fiber));
+        console.assert(dur >= 0);
+        // A backward ramp may not have an infinite duration (since it now has
+        // both a definite begin and end time).
+        console.assert(isFinite(dur) || p === 0);
+        f(p, fiber, this);
+        const begin = fiber.now - (isFinite(dur) ? p * dur : 0);
+        const observedDur = dur / fiber.rate;
+        const observedBegin = this.now - (isFinite(observedDur) ? p * observedDur : 0);
+        const observedEnd = observedBegin + observedDur;
+        this.ramps.set(fiber, { begin, dur, observedBegin, observedDur, observedEnd, f });
+        if (observedDur !== 0) {
+            this.scheduleFiber(fiber, Math.max(observedBegin, observedEnd));
+            return Fiber.Yield;
         }
     }
 
-    // Run all the ops for the fiber at this instant; that is, until the fiber
-    // yields (or ends).
-    runFiber(fiber) {
-        while (!fiber.yielded) {
-            if (fiber.rate > 0 && fiber.ip < fiber.ops.length) {
-                fiber.ops[fiber.ip++].call(fiber, this);
-            } else if (fiber.rate < 0 && fiber.unip > 0) {
-                fiber.unops[--fiber.unip].call(fiber, this);
-            } else {
-                break;
-            }
-        }
-        if (!fiber.yielded) {
-            this.fiberEnded(fiber);
-        }
+    // Remove the ramp object for the fiber and call its update function with
+    // the given p value (1 when going forward, 0 when backward). The ramp
+    // duration may have been cutoff, in which case the fiber local time should
+    // not be set to the past but match the current global time.
+    endRampForFiber(fiber, p) {
+        const { f, begin, dur, observedBegin, observedEnd } = this.ramps.get(fiber);
+        this.ramps.delete(fiber);
+        console.assert(isFinite(dur) || p === 0);
+        fiber.now = fiber.rate > 0 && observedEnd < this.now ?
+            (this.now - observedBegin) * fiber.rate : begin + (isFinite(dur) ? p * dur : 0);
+        f(p, fiber, this);
     }
 
-    // Execute all fibers scheduled in the [begin, end[ interval, then update
-    // all ramps.
+    // Change the duration of the current ramp (if any) for a fiber. If the
+    // new duration is less than the elapsed time, the fiber ends immediately.
+    // If the fiber is already at p = 1 (i.e., it is bound to be removed), this
+    // has no effect either. Setting the ramp duration for negative rates is
+    // not supported.
+    setRampDurationForFiber(fiber, dur) {
+        if (!this.ramps.has(fiber)) {
+            return;
+        }
+        const ramp = this.ramps.get(fiber);
+        if (dur === ramp.dur) {
+            return;
+        }
+        ramp.dur = dur;
+        if (fiber.rate === 0) {
+            // Let the observed values be handled when the rate changes again.
+            return;
+        }
+        ramp.observedDur = dur / fiber.rate;
+        console.assert(ramp.observedDur >= 0);
+        ramp.observedEnd = ramp.observedBegin + ramp.observedDur;
+        this.rescheduleFiber(fiber, Math.max(this.now, ramp.observedEnd));
+    }
+
+    // Run all fibers in the [begin, end[ time range.
     update(begin, end) {
         console.assert(this.instants.length === 0 || this.instants[0] >= begin);
         while (this.instants.length > 0 && this.instants[0] >= begin && this.instants[0] < end) {
@@ -381,35 +280,44 @@ export default class Scheduler {
             const queue = this.fibersByInstant.get(this.now);
             this.fibersByInstant.delete(this.now);
             while (queue.length > 0) {
-                const fiber = this.currentFiber = queue.shift();
-                console.assert(this.instantsByFiber.get(fiber) === this.now);
+                const fiber = queue.shift();
                 this.instantsByFiber.delete(fiber);
-                if (this.delays.has(fiber)) {
-                    const delay = this.delays.get(fiber);
-                    fiber.now = delay.fiberEnd;
-                    this.delays.delete(fiber);
+                this.spawnQueue = [];
+                this.joinQueue = [];
+                if (!this.fibers.has(fiber)) {
+                    this.fibers.set(fiber, fiber.run(this));
                 }
-                delete fiber.yielded;
-                fiber.lastUpdateTime = this.now;
-                this.resumeQueues = [[], []];
-                this.runFiber(fiber);
-                Array.prototype.unshift.apply(queue, this.resumeQueues[1]);
-                Array.prototype.unshift.apply(queue, this.resumeQueues[0]);
+                if (this.fibers.get(fiber).next().done) {
+                    this.fiberEnded(fiber);
+                }
+                Array.prototype.unshift.apply(queue, this.joinQueue);
+                Array.prototype.unshift.apply(queue, this.spawnQueue);
             }
         }
-        delete this.resumeQueues;
-        delete this.currentFiber;
+        delete this.spawnQueue;
+        delete this.joinQueue;
         this.now = end;
-        for (const [fiber, { delegate, begin, dur, rate, fiberBegin }] of this.ramps.entries()) {
-            if (!fiber.handleResult || fiber.rate === 0) {
+        for (const [fiber, ramp] of this.ramps.entries()) {
+            if (Object.hasOwn(ramp, "elapsed")) {
+                // Skip paused ramps.
                 continue;
             }
-            const p = (this.now - begin) / (dur / rate);
-            console.assert(p >= 0 && p <= 1);
-            if (p < 1) {
-                // The delegate is called with p = 1 when the ramp ends.
-                fiber.now = fiberBegin + p * dur;
-                delegate.rampDidProgress?.call(delegate, p, fiber, this);
+            if (isFinite(ramp.dur)) {
+                const p = (this.now - ramp.observedBegin) / ramp.observedDur;
+                console.assert(p >= 0 && p <= 1);
+                if (0 < p && p < 1) {
+                    // f is called with p=0 when the ramp begins, and p=1 when
+                    // the ramp ends. The local time is updated to reflect the
+                    // current progress of the ramp.
+                    fiber.now = ramp.begin + p * ramp.dur;
+                    ramp.f(p, fiber, this);
+                }
+            } else {
+                // Infinite ramps stay at p = 0, even when going backward (in
+                // which case it has a finite observed duration). Local time
+                // is still updated based on the observed elapsed time.
+                fiber.now = (this.now - ramp.observedBegin) * fiber.rate;
+                ramp.f(0, fiber, this);
             }
         }
         if (this.instants.length > 0) {
