+++ test/dodo.html
@@ -11,7 +11,7 @@
         <h1>Testin’ Dodo 🦤</h1>
         <div class="tests"></div>
         <p>
-            <a href="core.html">Testin’ Core Epistrophy</a> •
+            <a href="index.html">Testin’ Epistrophy</a> •
             <a href="../examples/">Examples</a>
         </p>
     </body>
