+++ test/index.html
@@ -11,7 +11,6 @@
         <h1>Testin’ Epistrophy</h1>
         <div class="tests"></div>
         <p>
-            <a href="core.html">Simpler core tests</a>&nbsp;&bull;
             <a href="dodo.html">Testin’ Dodo 🦤</a>&nbsp;&bull;
             <a href="../examples/">Examples</a>
         </p>
